[update] move file in source package[yuya]
