+++ app/src/main/java/com/schedule/shareships/data/repository/FirestoreRepository.kt
@@ -1,6 +1,6 @@
 package com.schedule.shareships.data.repository
 
-import com.schedule.shareships.data.SchedulePagingSource
+import com.schedule.shareships.data.source.SchedulePagingSource
 import com.schedule.shareships.model.Schedule
 
 interface FirestoreRepository {
